fix: Change to lazy evaluation of lambda function (as it should be)
interpretForm, when running a function, will check if it has the prop "lazyEval" set to true. if so, the args to the function are NOT evaluated. The function must handle this itself.

Since we are talking about language function here, evaluating the args actually means running the function. This means that when we set lazyEval, the function can run itself.

This could possibly be handled more generally, by always requiring the function to interpret itself. Will look into that.